Enabled projections for embedded resources in GET requests
To make this possible every SQL Alchemy mapped class got a new function _asdict() which returns the content of the class as a dict object. This is needed as Eve will fail to recursively add data when the data is implemented using SQL Alchemy. Instead the value of the projected field will be the returned dict object of the related class.