Added system for owners and affected users.
Every object can have an owner and an affected user, which is referenced through some id. This can also be indirect through a relation. The owner can perform all actions on that special object without permissions for that endpoint. The affected user can always read what affects him and can through a setting also get permission to DELETE or PATCH an object.

POST requests are parsed such that if the created object would belong to the posting user, access is granted. Therefore if the owner is stored in the object itself, everybody can post, effectivly creating a public resource!
This special behaviour is used for files.
In general the owner feature is used for forwards, which can be owned by the mailinglist owner.