Bugfixes concerning files
Fixes in several files:
In Schemas added a 'required: True' to the data of the file resource
In Models the images in joboffers and event were not created in the
database correctly
In Media the filestorage class was modified so it no longer breaks if a
resource contains no file and eve still checks for no file. Now returns
'None' correctly.