Custom file endpoint: Added test and varia
Renamed the file from download to file_endpoint to make the purpose
clearer.
Modified bootstrap to work with the new filename.
Added authentification test for the file endpoint.
Added a function 'new_file' in the test/util that allows creating a file
entry in the database (and creates a file in the storage folder as well)