Update to Eve 0.5.1 + Eve SQLAlchemy fork
Eve 0.5 moved SQLAlchemy support to a different package. We follow the
official eve repository now and use a fork of the extension. Changes are
only imports except for authentification.

The big change introduced by this is that the lookup query parameter now
only understands python syntax, but has support for embedding with it.

The fork introduces that values returned by Eve are now dict and
the SQLAResult class is removed of Eve-Sqlalchemy for now to resolve
problems with hooks, json and possibly other unexspected behaviour.

See issue #11 in eve-sqlalchemy
https://github.com/RedTurtle/eve-sqlalchemy/issues/11