Schema for read only fields
Cerberus supports setting fields to read-only so they can not be
provided manually.

Until now, the _author and title_id and desciption_id fields were
deleted so they wouldnt appear in the schema.

Now they are in the schema and correctly marked as read only
(Because it seems more approprieate than deleting them)