Validator: Eventsignups
All custom checks for eventsignups are not part of the validator

Removed item method patch for forwardusers because thats not needed
(was already removed for forwardaddresses, now correctly removed for both)

Some tests were implemented for patch, removed them.

One hook in validation.py converning eventsignups was covering
authorization. Moved it to authorization.py and updated bootstrap.

Also updated schema to include new validation.

(and  fixed a typo in settings)