Changed way identiy map is checked in test_tests
Until now we had a assertEqual(identityMap, {}).
For some reason this now fails,
since the identiy map is a WeakInstanceDict object. It is empty, but
not equal to a standart empty dict.

Maybe someone can find out why this test has stopped working, I really
could not find any clues.

Following some advice in [1] I changed it to a assertFalse to check if
the dict is empty, which works fine.

I am aware that this commit may fix just the symptoms and not causes,
so any comments are welcome.

[1]:
http://stackoverflow.com/questions/13312043/pythonefficient-way-to-check-if-dictionary-is-empty-or-not