Removed username, added email login
Closes #80

Instead of using the username for a session, it now uses the nethz.
(which is will be it almost all cases anyways)

Alternatively you can now login with email, but you have to explicitly
state so by using a "email" field instead of "nethz" field in your post to
sessions.

Updated tests and docs.

Also moved sessions test into auth test because they were doing the same.