Simplified post to /sessions
Removed the part that you have to specify 'nethz' in a 'nethz' field and
'email' in an 'email' field because it was kinda stupid.

Didn't really make the code for auth easier and any frontend tool would
require more effort (it would need to figure out wether to send the input
as nethz or email)

So I removed this again and put it in a general "user" field.

I did not go back to the old "username" because we don't have usernames
anymore.

Updated tests and docs.

Furthermore improved quality of post to sessions by using cerberus for
validation (More options e.g. no empty string for username)