Removed translatable fields in events and joboffers
Our functionalitz was nice an all but actually a little overkill which led to confusion when using the api.

And we dont have any situation where any other language than german and english.

Our solution would be for a usecase with many different languages.

So now we have a simplified solution which only includes german and english fields.

Also added a "catchphrase" field for events.