Fix bug that non-members could log in.
A POST to /sessions so far only checked if a user is in the database.
If you can get a session you are treated like a member and have access to files etc.

But actually your membership can be set to "none" and you should not be able to log in.

Added a test to show this behaviour and modified POST to /sessions to check membership status.

At the same time separated LDAP part of POST to /sessions better and improved logging.