LDAP overhaul.
Changed a lot of the current implementation to make it look more like a flask extension and make it easiert to use.

It is working together with the app now much better.
This required changes in cron -> made everything on cron work together with the app instead of initializing its own db connection.

Also ldap now handles importing users itself, adjusted authentication to use this. Its much clearer now, I hope.

Adjusted bootstrap for the new changes and ldap tests.