Group schema is now generated on demand.
Until now the group permission schema was generated at app creation.

This makes it a bit difficult to modularize the app since you have to
be very careful when to call it. Now its just generated when necessary which solves this problem.

Since its not very performance hungry this should be no big problem.