Testing db works now.
In the test settings the database name was stored in MONGO_DBTEST and Eve couldn't find it.
Changed to key to MONGO_DBNAME and now Eve uses the correct db for testing.

Also moved to test config into utils so its easier to find.

Furthermore removed the automatic database drop on test setup and replaced it with an
assert (that the db is empty) to make sure no database gets overwritten accidentially.