Restructured all tests
The tests are now in the directories of the respective module, e.g. all
users tests are in the /tests/users directory.

Tests can be run with
	py.test amivapi/tests

All not-updated tests have been moved to the "old tests" directory
Some of them might even work with the new version, but I haven't checked
that yet