Finished limited field of views without projection
Reason: Currently the projection is intended as a feature for user side
in eve. The projection in eve will always be overwritten by the request
and there are no easy ways to prohibit this.

So we now use a hook that will remove all hiden fields from the response
depending on the current user.

Performance wise, this is probably not optimal, but I think user request
are not happening so much that this would be an issue (unlike e.g. event
which will be accessed everytime someone opens the amiv page)