Review of new fixture system and test utils.
Its really nice! I just made some flake8 fixes regarding the comments
and modified a few of the for loops so the statements are easier to
read.

Furthermore the test config is now in the test class so it can be
modified by tests.

Mainly needed for ldap tests (coming soon (tm) )