Removed the root user from the database.
Like the anonymous user (already gone) it was needed mostly for database
relations in sql.
We don't really need this anymore.

Therefore I removed the root user from the database. To act as root you
don't need a session anymore, you can just send the root password as
token. The root user is not kind of a glorified api key.

This has the following advantages:
- We don't need to initialise the db anymore
- The users collection only contains actual people now

This means the tests don't need to create the root user and generally
the initdb function in mongo_manage is gone.