Fix for tests to work in windows
Programming on my brothers PC made me notice this.

All tests work fine, but in a MediaTest we open an nonexiting file. The
default mode is 'r' this seems to be a problem on windows since it
doesnt allow this mode for '' (whatever this means). But removing the
explicit 'r' lets python out what to use and it seems to work.