Simplify keyword handling in token parsing
While fixing flake8 errors I decided to simplify our token parsing altogether.
The primary crash was happening because of the special treatments of certain
keywords like "Token" and "Bearer", which, used on their own without an actual
token would cause a crash.

Now, the api does not treat any keyword special.
If a keyword is provided, it is ignored. This includes:
`Token <token>`, `Bearer <token>`, `Blablabla <token>`

We advocate not using a keyword anyway, so this maximizes compatibility.
Furthermore, since the special treatment of keywords is gone, the original
crash cannot happen anymore.