Allow zero-config api usage
By now, all configuration for API has useful default values or can handle
if values are unset (e.g. LDAP is deactivated if unconfigured)

Therefore it is not necessary to create a config before running the API.
`bootstrap.py` is changed a little to allow creating an app without config,
which is especially useful for tests and quick development servers.
Additionally, the configuration can now specified by environment vars, which
will be useful when working with Docker, which is planned for the near future.

All missing default values which were used by the CLI are moved into
`settings.py`.

The `create_config` command of the CLI is removed. In most cases, it is easier
to copy and modify an existing config.

The README is updated and now includes instructions on how to create a config
file (including config template for copypasting)