LDAP v3 (#250)
Until now, only the `ou` field form ldap was used to determine a students' department.

As it turns out, departments are never deleted from this list, i.e. if a student switches from ITET to INFK, both departments will be in the list, not only the current department.

We now use the `deparmentNumber` field to determine the department. It is a little bit more effort to parse, but at least we reliably know a students department now.