Bugfix: Missing `departmentNumber` in ldap should not crash the API
I was just importing missing users from LDAP and had to discover that the
ISG apparently is not that much into data validation, at least not when it
comes to LDAP.

A few (2, to be precise) users are missing the `departmentNumber` field
we need to determine the department.

In this case, the API should not crash but rather assume that the user
has no department.