Add Oauth2 login, remove python2
This commit mainly implements shiny new OAuth2 functionality to the API,
in form of a central login page which can be used for *Implicit Grants*.

Documentation for this new feature included.

Along this we decided to drop python 2 support to simplify the new
feature. This lead to cleaning up other parts of the code as well.

Finally, Travis and tox were adjusted to drop python2 as well.