Configure Eve to not merge nested documents
In it's default configuration, Eve merges nested documents on PATCH,
and therefore using PATCH, it is impossible to delete keys from nested
documents, e.g. it is not possible to remove single permissions from groups.

We patched Eve to allow turning this feature off.

With the switch to Eve 0.8, this feature is included in Eve and
so this patch can turn off merging of nested documents.

To make sure it's working, a test is added.

Resolves #267.