Fix test config and eventsignup validation
First of all, the test config would load sentry config from a file in the
current working dir. Not anymore :)

Furthermore, eventsignups with email would crash the API.

To fix this, a proper combination of `required` and `exclude` was needed,
which is able to to ignore None values. (None equals not set)

As Cerberus is a bit inconsequential about None values
(e.g. they can be ignored for `required`, but not for `exclude`), this
required patching how Cerberus handles exludes.

Resolves #213