Move schema description validation rules into general validator
We use additional schema rules such as 'description' to make our
schema more descriptive. However, Cerberus will complain about
unknown schema rules, so they have to be defined somewhere,
even if they don't do anything.

Currently, this is done in the documentation sub-module.
This is  a bit problematic, as our API is designed with replaceable
modules in mind -- but deactivating or replacing the doc module will
now cause errors in *all* schemas.

As the additional description rules are important whether or not we
use the online docs, I have moved the rules into the general validator.