Update settings, in particular version number, and serve logo from static files
Furthermore, I now ensured that the setup always uses the correct version
number.

This required me to remove the 'hack' to always install the current folder
from `requirements.txt`. This was quite unusual anyways.

As a result tox etc. needed to be adjusted.

Additionally, the logo used for OAuth and (soon) docs is now served consistently from a static endpoint.