Fix validator crashes
I have noticed that Cerberus swallows exceptions.
I modified the test framework to detect this.

This showed two errors in our validators. One was a missing
return statement, the other was more fundamental:

Whenever a file validator would register an error, Cerberus
would crash, as it internally saved the validated value
and for some reason deepcopies it during the processing.
File buffers couldn't be deepcopied, causing a crash.

I have solved this by removed the reference to the validated
value, as it is *never* used anyways.
(I have no clue why it is in there in the first place)