Fix token in email (#366)
TherThe email tests failed because the token could not be decoded. This PR fixes this.

The tokens were returned as bytes, and for some reason, the url_for that turns the token into
a clickable link would take the b'...' and apply some escaping, resulting in b%27...%27.
I now switched to the url-safe version of the signer, and everything is well.

I'm not quite sure why the tests did not fail before, but now everything is working as intended.