Allow proper caching with `If-Modified-Since`.
There is a subtle difference between `no-store` and `no-cache`.

`no-store`: Never return from cache, always request new. (Especially problematic for images, which are always re-loaded. *This is what we currently have*.

`no-cache`: Don't return from cache by default, but first re-validate with the server, in our case using the `If-Modified-Since` header, which browsers use automatically, as eve includes `Last-Modified` headers in all responses. *This is what we actually want.*